Addition of 2 side projects
Added 2 side projects to my portfolio that i want to begin development on. The Good Habit tracker which will be a simple desktop application and the Script tracker which will be an website that uses Google APIs to pull information from a google sheet that I've been maintaining of the years.